feat: split handler.go to multiple handlers
